requirements.txt
ollama 
requests
beautifulsoup4
tavily-python
python-dotenv
